fix(repo): rename target project for tests
chore(repo): move webhook tests to own file
